+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/MemoryCache.cs
